updated UI for dex scroll
