+++ mobile-app/app/screens/AppNavigator/screens/Dex/DexScreen.tsx
@@ -12,7 +12,7 @@ import {
   SkeletonLoader,
   SkeletonLoaderScreen
 } from '@components/SkeletonLoader'
-import { ThemedIcon, ThemedScrollViewV2, ThemedTextV2, ThemedTouchableOpacityV2, ThemedViewV2 } from '@components/themed'
+import { ThemedScrollViewV2, ThemedTextV2, ThemedViewV2 } from '@components/themed'
 import { tailwind } from '@tailwind'
 import { translate } from '@translations'
 import { DexParamList } from './DexNavigator'
@@ -22,11 +22,7 @@ import { EmptyActivePoolpair } from './components/EmptyActivePoolPair'
 import { debounce } from 'lodash'
 import { ButtonGroupTabKey, PoolPairCards } from './components/PoolPairCards/PoolPairCards'
 import { ButtonGroupV2 } from './components/ButtonGroupV2'
-import { ScrollView } from 'react-native'
-import { AssetsFilterItem } from '../Portfolio/components/AssetsFilterRow'
 import { HeaderSearchInputV2 } from '@components/HeaderSearchInputV2'
-import { DexScrollable } from '@screens/AppNavigator/screens/Dex/components/DexScrollable'
-import { TotalValueLocked } from './components/TotalValueLocked'
 
 enum TabKey {
   YourPoolPair = 'YOUR_POOL_PAIRS',
@@ -43,30 +39,6 @@ export function DexScreen (): JSX.Element {
   const [activeTab, setActiveTab] = useState<string>(TabKey.AvailablePoolPair)
   const tokens = useSelector((state: RootState) => tokensSelector(state.wallet))
   const [expandedCardIds, setExpandedCardIds] = useState<string[]>([])
-  const { tvl } = useSelector((state: RootState) => state.block)
-
-  const buttonGroup = [
-    {
-      id: ButtonGroupTabKey.AllPairs,
-      label: translate('screens/DexScreen', 'All pairs'),
-      handleOnPress: () => onButtonGroupChange(ButtonGroupTabKey.AllPairs)
-    },
-    {
-      id: ButtonGroupTabKey.DFIPairs,
-      label: translate('screens/DexScreen', 'DFI pairs'),
-      handleOnPress: () => onButtonGroupChange(ButtonGroupTabKey.DFIPairs)
-    },
-    {
-      id: ButtonGroupTabKey.DUSDPairs,
-      label: translate('screens/DexScreen', 'DUSD pairs'),
-      handleOnPress: () => onButtonGroupChange(ButtonGroupTabKey.DUSDPairs)
-    },
-    {
-      id: ButtonGroupTabKey.FavouritePairs,
-      label: translate('screens/DexScreen', 'Favourites'),
-      handleOnPress: () => onButtonGroupChange(ButtonGroupTabKey.FavouritePairs)
-    }
-  ]
 
   const onButtonGroupChange = (buttonGroupTabKey: ButtonGroupTabKey): void => {
     setExpandedCardIds([])
@@ -89,6 +61,7 @@ export function DexScreen (): JSX.Element {
       }))
     return _yourLPTokens
   })
+
   const onTabChange = (tabKey: TabKey): void => {
     setActiveTab(tabKey)
   }
@@ -242,6 +215,11 @@ export function DexScreen (): JSX.Element {
     }
   }, [pairs])
 
+  const onSearchBtnPress = (): void => {
+    setShowSearchInput(true)
+    setFilteredAvailablePairs([])
+  }
+
   // Top Liquidity pairs
   const [topLiquidityPairs, setTopLiquidityPairs] = useState<Array<DexItem<PoolPairData>>>(pairs)
   useEffect(() => {
@@ -285,66 +263,21 @@ export function DexScreen (): JSX.Element {
           </ThemedTextV2>
         </View>)
       : (
-        <>
-          <ThemedViewV2
-            light={tailwind('bg-mono-light-v2-00 border-mono-light-v2-100')}
-            dark={tailwind('bg-mono-dark-v2-00 border-mono-dark-v2-100')}
-            style={tailwind('flex flex-col items-center pt-4 rounded-b-2xl border-b')}
-          >
-            <View style={tailwind('w-full px-5')}>
-              <ButtonGroupV2
-                buttons={tabsList}
-                activeButtonGroupItem={activeTab}
-                testID='dex_tabs'
-                lightThemeStyle={tailwind('bg-transparent')}
-                darkThemeStyle={tailwind('bg-transparent')}
-              />
-            </View>
-          </ThemedViewV2>
-          <View style={tailwind('my-4')}>
-            <ThemedViewV2 testID='dex_button_group'>
-              <ScrollView
-                horizontal
-                showsHorizontalScrollIndicator={false}
-                contentContainerStyle={tailwind('flex justify-between items-center flex-row px-5')}
-              >
-                <ThemedTouchableOpacityV2
-                  onPress={() => {
-                    setShowSearchInput(true)
-                    setFilteredAvailablePairs([])
-                  }}
-                  style={tailwind('text-center pr-4')}
-                  testID='dex_search_icon'
-                >
-                  <ThemedIcon
-                    iconType='Feather'
-                    name='search'
-                    size={24}
-                    light={tailwind('text-mono-light-v2-700')}
-                    dark={tailwind('text-mono-dark-v2-700')}
-                  />
-                </ThemedTouchableOpacityV2>
-                {buttonGroup.map((button, index) => (
-                  <AssetsFilterItem
-                    key={button.id}
-                    label={button.label}
-                    onPress={button.handleOnPress}
-                    isActive={activeButtonGroup === button.id}
-                    testID={`dex_button_group${button.id}`}
-                    additionalStyles={!(buttonGroup.length === index) ? tailwind('mr-3') : undefined}
-                  />
-                  )
-                )}
-              </ScrollView>
-            </ThemedViewV2>
+        <ThemedViewV2
+          light={tailwind('bg-mono-light-v2-00 border-mono-light-v2-100')}
+          dark={tailwind('bg-mono-dark-v2-00 border-mono-dark-v2-100')}
+          style={tailwind('flex flex-col items-center pt-4 rounded-b-2xl border-b')}
+        >
+          <View style={tailwind('w-full px-5')}>
+            <ButtonGroupV2
+              buttons={tabsList}
+              activeButtonGroupItem={activeTab}
+              testID='dex_tabs'
+              lightThemeStyle={tailwind('bg-transparent')}
+              darkThemeStyle={tailwind('bg-transparent')}
+            />
           </View>
-          {activeTab === TabKey.AvailablePoolPair && !showSearchInput && (
-            <TotalValueLocked tvl={tvl ?? 0} />
-          )}
-          <TopLiquiditySection onPress={onSwap} pairs={topLiquidityPairs} />
-          <NewPoolsSection onPress={onAdd} pairs={newPoolsPairs} />
-
-        </>
+        </ThemedViewV2>
       )}
       <View style={tailwind('flex-1')}>
         {activeTab === TabKey.AvailablePoolPair &&
@@ -368,6 +301,11 @@ export function DexScreen (): JSX.Element {
               setIsSearching={setIsSearching}
               searchString={searchString}
               showSearchInput={showSearchInput}
+              onButtonGroupChange={onButtonGroupChange}
+              activeButtonGroup={activeButtonGroup}
+              onSearchBtnPress={onSearchBtnPress}
+              topLiquidityPairs={topLiquidityPairs}
+              newPoolsPairs={newPoolsPairs}
             />
           )}
 
@@ -387,51 +325,14 @@ export function DexScreen (): JSX.Element {
             setIsSearching={setIsSearching}
             searchString={searchString}
             showSearchInput={showSearchInput}
+            onButtonGroupChange={onButtonGroupChange}
+            activeButtonGroup={activeButtonGroup}
+            onSearchBtnPress={onSearchBtnPress}
+            topLiquidityPairs={topLiquidityPairs}
+            newPoolsPairs={newPoolsPairs}
           />
         )}
       </View>
     </ThemedViewV2>
-  )
-}
-
-function TopLiquiditySection ({ pairs, onPress }: {pairs: Array<DexItem<PoolPairData>>, onPress: (data: PoolPairData) => void}): JSX.Element {
-  return (
-    <DexScrollable
-      testID='dex_top_liquidity'
-      sectionHeading='TOP LIQUIDITY'
-      sectionStyle={tailwind('my-6')}
-    >
-      {pairs.map((pairItem, index) => (
-        <DexScrollable.Card
-          key={`${pairItem.data.id}_${index}`}
-          poolpair={pairItem.data}
-          style={tailwind('mr-2')}
-          onPress={() => onPress(pairItem.data)}
-          label={translate('screens/DexScreen', 'Swap')}
-          testID={`composite_swap_${pairItem.data.id}`}
-        />
-      ))}
-    </DexScrollable>
-  )
-}
-
-function NewPoolsSection ({ pairs, onPress }: {pairs: Array<DexItem<PoolPairData | WalletToken>>, onPress: (data: PoolPairData, info: WalletToken) => void}): JSX.Element {
-  return (
-    <DexScrollable
-      testID='dex_new_pools'
-      sectionHeading='NEW POOLS'
-      sectionStyle={tailwind('mb-6')}
-    >
-      {pairs.map((pairItem, index) => (
-        <DexScrollable.Card
-          key={`${pairItem.data.id}_${index}`}
-          poolpair={pairItem.data as PoolPairData}
-          style={tailwind('mr-2')}
-          onPress={() => onPress(pairItem.data as PoolPairData, pairItem.data as WalletToken)}
-          label={translate('screens/DexScreen', 'Add to LP')}
-          testID={`add_liquidity_${pairItem.data.id}`}
-        />
-      ))}
-    </DexScrollable>
   )
 }
